+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+* Moved chicken food items to item tag `iguanatweaksreborn:chicken_food_items`
+  * Also fixed not being able to use any item on chickens
+
 ## 3.5.2
 * Players are no longer able to pillar **only if** there are monsters nearby
 * Instant heal food now heal more